Conflicts:
	code/_onclick/hud/human.dm

+++ code/modules/power/gravitygenerator.dm
@@ -36,9 +36,10 @@ var/const/GRAV_NEEDS_WRENCH = 3
 	if(prob(20))
 		set_broken()
 
-/obj/machinery/gravity_generator/tesla_act(var/power)
+/obj/machinery/gravity_generator/tesla_act(power, explosive)
 	..()
-	qdel(src)//like the singulo, tesla deletes it. stops it from exploding over and over
+	if(explosive)
+		qdel(src)//like the singulo, tesla deletes it. stops it from exploding over and over
 
 /obj/machinery/gravity_generator/update_icon()
 	..()
